refactor:  extract schema for category and the middleware function validateRequest
